more devices, and some architectural bikeshedding
arghhhh why is ownership so haaaaaard. I want to do things The Right Way
but I'm not sure how :(
Hopefully, the right architecture will reveal itself as I get more
familiar with the system. As such, I'm just going to push forward, and
refactor things later.

Oh, and the code actually makes it to the LCD code! Time to hack
together a LCD device and get some video output :D