enable using flash ROM dump + misc tweaks
turns out there was a lil bug in ipodloader 1 that I needed to fix as
well. it tries to detect what model ipod it's running on by reading a
u32 from flash ROM and matching it to a string as follows:

`if (inl(0x2000) == (unsigned)"gfCS") { ... }`

that's not correct, as instead of casting the contents of the static
string as a unisigned, it's casting the char* pointer itself to a
unsigned!

changing the line to:

`if (inl(0x2000) == *(unsigned*)"gfCS") { ... }`

fixes it right up.