add usec timer + untested irq-capable timers
I straight up copied the timer device from ts7200, and tweaked it enough
so that it _should_ work with the iPod. I have no way of testing it at
the moment, and odds are it _doesn't_ do what's expected (since I just
pulled a random KHz frequency out of my ass lol), but hey, the
foundation is there.

Next up on the "Unimplemented Memory" block is GPIO...

* * *

Organization is a bit of a mess at the moment, since I decided to make a
"timers" bus device, which overlaps with the memory allotted to the
Syscon by the memory map. Things still work, but it's really brittle.

Now that I think about it, having these "groups" of devices as their own
separate bus devices is dumb. The only real reason they exist is to make
the memory map shorter (which is silly). I'll have to do something about
that (maybe a solution will present itself when I rewrite the current
memory system?)...