more stub devices + inject HLE sysinfo to fastram
the sysinfo code is really hacky, but it seems to work. I spent quite a
bit of time poking around in Ghidra to try and figure out where the
flashROM bootloader writes this info to, but in the end, I didn't figure
it out. That said, I really enjoyed playing with Ghidra, and I think
it'll prove invaluable as a I keep working on the project!