add IDE standby command - can load RetailOS!!!
and with that, we can actually get ipodloader2 to boot into the apple
firmware from disk! holy shit!

I'm definately going to take a step back here and work on un-stubbing a
lot of stuff, with emphasis on the keypad. This will involve finally
moving the UI thread from the LCD controller to the top-level.

I also want to look into how to speed up ATA reads, because right now,
that retailOS progress bar is sloooooooooooooooooow as heck! Maybe I can
do some BufferedReader shenanigans to get some performance...