add support for using a Flash ROM dump + cold boot
I feel like enough of the system is implemented that I *might* be able
to get a cold-boot up and running... Right now, it crashes trying to
access the cache controller, but who knows, maybe if I stub it out,
it'll Just Work out of the box lol (probably not tho).

This commit also gave me an excuse to finally clean up the rat's nest
that was `Ipod4g::new()`. I finally separated the core system init code
from the HLE bootloader, and neatly tucked all the HLE bootloader stuff
into it's own module. Hooray for code cleanup!