many new stubs + LCD controller improvements
This is the result of trying to see how far I could get by cold-booting
from the Flash ROM. The answer: pretty far! Right now, it's caught in an
endless loop related to I2C stuff, and I haven't had a chance to explore
what's the cause of the loop. If it's like some of the other stuff I've
encountered, it might be as simple as asserting a bit in one of the I2C
registers (though given the more complex bus-like nature, it might not
be that simple... we'll see I guess)

* * *

Can you tell I'm procrastinating on plumbing through the controls?