more flash ROM bootloader related stubs/patches
it looks like the first time a firmware image is flashed onto a HDD, the
iPod will try to run the FlashROM update program. It can tell that if
it's been run on not based on the calue of the `id` bit under the aupd
image. If it's zero, it'll run the update, but if it's one, it'll skip
it.

I wrote a little script called `patch_aupd_bit.sh` which forcefully
toggles this bit on. This _should_ let me skip implementing the FlashROM
update while still leveraging the LLE FlashROM bootloader.

oh, and leaving the bit set to 1 shows some _very_ fun error images!
It's some honest-to-goodness closed-source Apple code running in the
emulator! Very cool!