move pp-specific devices to their own module
while the surface-level goal of this project is to get PP50xx devices up
and running, I _do_ have a loftier goal of using this project as a
foundation for a QEMU-like application that supports a whole bunch of
platforms. These other platforms won't be supported for a _while_, but
hey, might as well keep the project condusive to future expansion,
right?