create device prelude module
I've been meaning to do this for a while. Quite often, I would end up
acting on the clippy lint to remove a log::Level/BitField import, only
to re-import it after realizing I need it a few minutes later.

This prelude will cut down on boilerplate, and give me quick and easy
access to key traits that devices rely on.