update to gdbstub 0.2
Ah, it's good to be back. Working on `gdbstub` was a fun diversion, but
it'll be fun to get back into the swing of things with `clicky`!

The killer feature of `gdbstub` 0.2 (at least for clicky) is
multithreading support, as now I'll be able to properly debug code that
runs on the coprocessor. How exciting!

That said, before diving into implementing more hardware, I think I'll
take some time to look at the clicky codebase with "fresh eyes", since
I've been away from the project for a few weeks now, and I might be able
to spot some improvements I wouldn't have thought of earlier.