start on Rockbox support
One key thing I've found is that `ipodloader2` doesn't seem to be able
to load Rockbox correctly. From what I can tell, it overwrites the data
at 0x1000_0000 _before_ the COP has been woken up, which results in the
COP starting execution smack in the middle of the Rockbox bootloader!

I'm curious if this is the result of a bug somewhere in my code, or if
that's what happens on actual hardware, and it just so happens that the
control flow eventually "fixes" itself.

Ah well, either way, I've switched gears and decided to use Rockbox's
built-in bootloader instead.

I've added some kinda vague instructions to the README on how to build
Rockbox (mostly just so I don't forget), and real-talk, it's kinda jank.
Having to use the make_fw utility seems kinda shitty. Maybe I should
look into patching `ipodpatcher` to work with disk images directly?

* * *

I've managed to get the Rockbox firmare booting, and as it stands, it
dies performing an unimplemented access on a Timer device (which is
totally fine!). To get to this point, I've had to tweak / add some new
features to various devices:

- (memcon) added more hard-coded mappings
- (memcon) per-CPU memory mappings
- (ide) stubbed support for multi-sector reads
- (cpucon) removed COP only-enable-once hack + lil refactoring
- (devcon) a little stubbyboi for the cache priority (whatever that it)

* * *

There are two different approaches I can take to tackle the timer
device:

1. Just spin up a dedicated thread for timer interrupts, and don't worry
too much about it.
2. spend some time doing the BIG THINK about how to handle asynchrony in
the codebase as a whole, potentially re-architecting things to support
async calls in devices.

That latter option is something I've been thinking about for the past
little while. The first time the issue of asynchrony reared its head was
while I was working on the IDE device. Reading blocks really aught to be
an asynchronous operation, but I was lazy, and decided to block the main
thread when a read/write request came in. Well, now I have to implement
timers, and blocking isn't really an option... Plus, i'll have to
implement DMA at some point as well, so....

I'd like to avoid using threads directly, as one of my "would be nice"
goals is porting `clicky` to WebAssembly, and running it in the browser,
and at the time of writing, WebAssembly doesn't support threads.

The other approach that I'd like to explore is writing a custom async
executor! Basically, I'd want to be able to spawn tasks from devices,
and drive them forward between CPU cycles. I'm leaning towards
rolling-my-own executor since full-system-emulation isn't really a
use-case most existing executors are optimized for, and I'm worried
about overhead. Plus, I vaguelly recall that most executors use a
dedicated thread from I/O and/or timers, which clashes with my wish
above...