add IRQ support to cfg_timer
As I talked about in 8d8ab33, there's going to be quite a bit of effort
to handle asynchrony in an ergonomic and performant manner in `clicky`.

While I definately think an async/await + custom executor approach is
the way to go, I've decided to kick the can down the road a little while
longer, and use threads for now. Each timer spawns a worker thread which
runs an "interrupter" task. This closely parallels the implementation in
ts7200.

**That being said,** I am taking a forward-looking approach here, and
decided to write the timer workers as async/await tasks. The thread
simply uses the `block_on` executor to run the worker task.

With this approach, it _should_ be relatively straightforward to
eventually swap out the `thread::spawn + block_on` implementation with
a proper executor `spawn` based implementation.

* * *

Another thing I forgot to mention: It'd be nice to implement some sort
of "save-state" functionality, where the entire system's state can be
saved/restored. Prior to this commit, this actually would have been
fairly straightforward, since there weren't any background tasks. I
don't know a great way to serialize the state of an asynchronous system,
but I thought I'd put this thought down in writing just so it's
something I keep in mind moving forward.