use Async IO for block devices
The IDE drive still uses block_on to run the async operations, but it
should be possible to offload these operations onto an executor at some
point...

I spent quite a bit of time futzing around with different architectures
to do things asynchronously in the IDE drive, and truth be told, I
couldn't get something I was happy with up and running!

One solution resulted in wrapping a bunch of internal IDE state in an
ArC<Mutex<_>>, which was kinda ugly. Another experiment involved using
something similar to the timer device's "interrupter" thread, where a
separate task would do IO ops for the main IDE device, and communicate
with the device over an async_channel. This sort-of worked, but I ran
into some issues around code cleanliness and readability...

If I was under some sort of deadline, I'd probably just push what I came
up with and cleaned it up later, but hey, this is my side-project, so I
can just _not_ do that and leave things as-is for a little while!

I'll probably revisit this problem later on, potentially when the time
comes to implement DMA. I have a feeling that the asynchronous nature of
DMA will lead to some substantial architectural changes, which might
shed some light on how to implement IDE drive asynchrony properly...