fix hang on Rockbox splash screen
Ughhhhh this was an annoying bug to track down. I'm starting to reach
the point in the project where I'm not "failing loudly" anymore (e.g: by
not having a device implemented / hitting unmapped memory), but instead,
the code is ostensibly running correctly, but my _implementation_ has a
bug that's causing it to hang...

This particular hang exposed a couple of different issues in my IRQ
implementation:
1. I needed to implement Mailbox IRQs, which ended up requiring some
changes to the interrupt controller, as unlike every other IRQ in the
system, mailbox IRQs are core-specific!
2. My premature efficiency trick with `irq_pending` bit me in the ass,
since it didn't account for the possiblity of CPU cores _internally_
disabling interrupts... I ended up half-assing a fix, but I really aught
to take a closer look.
3. So, uh, turns out cloning signal::Master had a bug in it lol. The
addition of an Arc fixed it right up.

I really aught to revisit my approach to IRQs... It's kinda hard to
follow tbh.