reduce log spam when sitting in Rockbox menu
I think I've got the cpucon and mailboxes working pretty much correctly
now, so there's no need to spam the logs with stub messages and debug
info.

I also added a toggle to the AsanRAM so that it only reports
uninitialized memory access once, and then marks it as initialized.
Rockbox likes to use a CRC function to validate metadata blocks in it's
slab allocator by casting it's struct pointer to a char*. Unfortunately,
the compiler isn't required to initialize any padding bytes in the
structure, which means lots of errors get thrown when linearly iterating
through the struct as though it was a byte array.

With this change, the only log spam during Rockbox's steady-state menu
operation is the occational poll of the i2c registers.