fix time on wasm + use executor instead of threads!
Wowie, time was annoying to fix... It turns out async_timers had a bug
in it, so I had to write a new crate (relativity) to get the
platform-agnostic timer working properly (using gloo-timers for wasm). I
should probably open an issue on async_timers though, since their wasm
implementation is broken...

Also, look at this shiny new executor system! It uses multiple threads
on the desktop, and a single thread on wasm, and it all Just Works!
Neato!