major i2c overhaul + working RTC!
wow, this was a real yak-shave.

All I wanted to do was get the RTC working, but as it turns out, the RTC
is located on the iPod's power management controller, which is connected
via the i2c bus.

As such, I've had to un-stub the i2c controller, and actually implement
it properly. I think it turned out mostly okay, but boy-oh-boy, was
getting it working annoying:

- Translating / Logging `MemResult`s from the i2c bus is kinda janky,
and took quite a bit of experimenting to even get working.
- I haven't ever used i2c before, so I had to do a _lot_ of reading up
on how the protocol is supposed to work. From what I understand, the
core i2c protocol only supports reading/writing bytes to a device (given
by it's 7-bot address), but doesn't say anything on how to read/write
registers from individual devices. This is kinda annoying, since i've
had to include the required latching-logic on the device side...

Regardless of the rocky journey, the end result is that I now have a
clean little i2c interface implementation, and I can work on the more
interesting parts of the problem: actually implementing the devices
connected to the i2c bus.

The next order of business is stubbing out / implementing a couple more
registers on the pcf5060x device.

* * *

Additionally, I've split the clickwheel controller into a separate
device (called opto). I'm kinda confused what the hell is going on,
since from the looks of it, the opto + keypad doesn't use any of the i2c
architecture that other devices use, and kinda just does it's own
thing. I mean, I guess there's _some_ precedence for these sorts of
shenanigans (like how there are 2 separate DMA controllers on the board,
one of which is dedicated to ATA HDDs), but still, weird.