ide: Use Spawner for updating registers, all operation delay
I recently noticed that iPodLinux was struggling to boot due to the IDE
IRQ handling being set right after sending IDE commands.

This PR introduces a `task_spawner` member to IDE driver for deferring
operation. It also adds basic implementations of Recal/ReadNativeMax.

A significant amount of code is due to the use of Arc<RwLock<IdeRegs>>
for thread safety.

This PR allows iPodLinux hda initialization to go a little further, but
work in ReadSectors still needs to be done.